requirements.txt
chromedriver_autoinstaller==0.6.4
discord.py==2.2.3
selenium==4.17.2
